+++ client/static/scripts/data_collection/GamePiece.js
